implemented shortcut version
